move and rename class
